feat: add real databases for order handlers

+++ common-lib/src/main/java/ru/kaznacheev/restaurant/common/exception/OrderNotFoundException.java
@@ -6,12 +6,14 @@ import ru.kaznacheev.restaurant.common.dto.response.ResponseTitle;
  * Исключение, выбрасываемое при попытке получения несуществующего заказа.
  */
 public class OrderNotFoundException extends BaseException {
+
     /**
-     * Создает экземпляр класса {@link OrderNotFoundException} с указанными значениями для {@link BaseException}.
+     * Конструктор исключения.
      *
      * @param id Идентификатор заказа, который вызвал исключение
      */
-    public OrderNotFoundException(int id) {
+    public OrderNotFoundException(Long id) {
         super(ResponseTitle.NOT_FOUND.name(), ResponseTitle.NOT_FOUND.getStatus(), "Заказ с идентификатором " + id + " не найден");
     }
+
 }
